split monitor and update as they take hella long
