+++ pkg/update/update.go
@@ -1,4 +1,4 @@
-package updatemonitor
+package update
 
 import (
 	"crypto/sha256"
@@ -21,7 +21,7 @@ func getContent(arr []byte) (res seq[byte]) {
 @*/
 
 // ==================================================================================
-// ============================= Update & Monitor Types =============================
+// ============================= Update Types =======================================
 
 type UpdateResponse struct {
 	Full_tree_head client.FullTreeHead
@@ -48,20 +48,6 @@ pred (u UpdateResponse) Inv() {
 	u.Prev_search.Inv() &&
 	acc(u.Values) &&
 	acc(u.Openings)
-}
-@*/
-
-type MonitorResponse struct {
-	Full_tree_head client.FullTreeHead
-	Binary_ladder  []proofs.BinaryLadderStep
-	Search         proofs.CombinedTreeProof
-}
-
-/*@
-pred (m MonitorResponse) Inv() {
-	m.Full_tree_head.Inv() &&
-	acc(m.Binary_ladder) &&
-	m.Search.Inv()
 }
 @*/
 
@@ -239,38 +225,6 @@ func buildUpdatePrefixTrees(resp UpdateResponse, n int /*@, ghost p perm @*/) (t
 	return trees, rootHashes, nil
 }
 
-// buildMonitorPrefixTrees constructs prefix trees from a MonitorResponse's prefix proofs.
-// @ requires noPerm < p
-// @ requires acc(resp.Inv(), p)
-// @ ensures err == nil ==> acc(resp.Inv(), p)
-// @ ensures err == nil ==> acc(trees) && len(trees) == n
-// @ ensures err == nil ==> forall j int :: 0 <= j && j < n ==> acc(&trees[j])
-// @ ensures err == nil ==> forall j int :: 0 <= j && j < n ==> trees[j] != nil
-// @ ensures err == nil ==> forall j int :: {&trees[j]} 0 <= j && j < n ==> trees[j].Inv()
-// @ ensures err == nil ==> acc(rootHashes) && len(rootHashes) == n
-// @ ensures err == nil ==> forall j int :: 0 <= j && j < n ==> acc(&rootHashes[j])
-// @ ensures err == nil ==> forall j int :: 0 <= j && j < n ==> acc(rootHashes[j])
-// @ ensures err == nil ==> forall j int :: 0 <= j && j < n ==> forall k int :: 0 <= k && k < sha256.Size ==> low(rootHashes[j][k])
-// @ ensures err != nil ==> acc(resp.Inv(), p)
-// @ trusted
-func buildMonitorPrefixTrees(resp MonitorResponse, n int /*@, ghost p perm @*/) (trees []*proofs.PrefixTree, rootHashes []*[sha256.Size]byte, err error) {
-	trees = make([]*proofs.PrefixTree, 0, n)
-	rootHashes = make([]*[sha256.Size]byte, 0, n)
-	//@ unfold acc(resp.Inv(), p)
-	for i := 0; i < n; i++ {
-		prf := /*@ unfolding acc(resp.Search.Inv(), p/2) in @*/ resp.Search.Prefix_proofs[i]
-		if tree, treeErr := prf.ToTree(resp.Binary_ladder); treeErr != nil {
-			//@ fold acc(resp.Inv(), p)
-			return nil, nil, treeErr
-		} else {
-			trees = append( /*@ perm(1/2), @*/ trees, tree)
-			rootHashes = append( /*@ perm(1/2), @*/ rootHashes, tree.Value)
-		}
-	}
-	//@ fold acc(resp.Inv(), p)
-	return trees, rootHashes, nil
-}
-
 // ==================================================================================
 // ============================= VerifyUpdate =======================================
 
@@ -416,232 +370,68 @@ func VerifyUpdate(st *client.UserState, label []byte, resp UpdateResponse, confi
 // @ requires acc(config, p)
 // @ requires resp.Full_tree_head.Tree_head.Tree_size > 0
 // @ requires resp.Full_tree_head.Tree_head.Tree_size <= uint64(len(resp.Search.Prefix_proofs))
+// @ requires low(resp.Full_tree_head.Tree_head.Tree_size)
+// @ requires low(resp.New_version)
 // @ ensures acc(label, p) && acc(config, p)
 // @ ensures acc(resp.Inv(), p)
-// @ trusted
 func VerifyHistory(label []byte, resp UpdateResponse,
 	config *client.Configuration /*@, ghost p perm @*/) (err error) {
 
-	size := resp.Full_tree_head.Tree_head.Tree_size
-	n := len(resp.Search.Prefix_proofs)
-
-	// Build prefix trees from current tree proof
-	trees, rootHashes, buildErr := buildUpdatePrefixTrees(resp, n /*@, p @*/)
-	if buildErr != nil {
-		return buildErr
-	}
-
-	// Determine version range to check
-	start := uint32(0)
-	if resp.Prev_greatest != nil {
-		start = *resp.Prev_greatest + 1
-	}
-
-	// Verify each new version exists in the current log
-	for v := start; v <= resp.New_version; v++ {
-		prev := uint32(0)
-		if v > 0 {
-			prev = v - 1
-		}
-		ok, verifyErr := VerifyUpdateKey(trees, rootHashes, size, label, v, prev, config /*@, p/4 @*/)
-		if verifyErr != nil {
-			return verifyErr
-		}
-		if !ok {
-			return errors.New("history verification failed: version not found in log")
-		}
-	}
-	return nil
-}
-
-// ==================================================================================
-// ============================= VerifyMonitor ======================================
-
-// VerifyMonitor verifies that previously-searched label-version pairs remain in the
-// log (Section 8.2 — Contact Monitoring).
-// Uses CheckGreatest on all frontier nodes. For monitoring, res == -1 (hole) is a
-// failure, while res == 0 or 1 are acceptable (version exists in log).
-// The low(Version) property is preserved from the input monitoring map (established
-// by prior VerifyLatest/VerifyUpdate calls), not re-proven via TStar.
-
-// @ requires noPerm < p
-// @ preserves st.Inv()
-// @ requires acc(resp.Inv(), p)
-// @ requires acc(label, p)
-// @ requires low(len(label)) && (forall i int :: 0 <= i && i < len(label) ==> low(label[i]))
-// @ requires acc(monitor_map)
-// @ requires acc(config, p)
-// @ requires low(resp.Full_tree_head.Tree_head.Tree_size)
-// @ requires resp.Full_tree_head.Tree_head.Tree_size > 0
-// @ requires resp.Full_tree_head.Tree_head.Tree_size <= uint64(len(resp.Search.Prefix_proofs))
-// @ requires resp.Full_tree_head.RootHash != nil
-// @ requires low(len(monitor_map))
-// @ requires forall j int :: 0 <= j && j < len(monitor_map) ==> low(monitor_map[j].Version)
-// @ requires forall j int :: 0 <= j && j < len(monitor_map) ==> low(monitor_map[j].Position)
-// @ requires low(config.ReasonableMonitoringWindow)
-// @ ensures acc(resp.Inv(), p) && acc(label, p) && acc(config, p)
-// @ ensures acc(new_map)
-// @ ensures err == nil ==> low(len(new_map))
-// @ ensures err == nil ==> forall j int :: 0 <= j && j < len(new_map) ==> low(new_map[j].Version)
-// @ ensures err == nil ==> forall j int :: 0 <= j && j < len(new_map) ==> low(new_map[j].Position)
-func VerifyMonitor(st *client.UserState, label []byte, resp MonitorResponse, monitor_map []client.MonitoringMapEntry, config *client.Configuration /*@, ghost p perm @*/) (new_map []client.MonitoringMapEntry, err error) {
 	//@ unfold acc(resp.Inv(), p)
 
 	determined := false
 	var resultErr error = nil
 
-	// Capture size and n before folding so Gobra retains size <= n
 	size := resp.Full_tree_head.Tree_head.Tree_size
 	n := len(resp.Search.Prefix_proofs)
 
-	// Phase 1: UpdateView
-	updateErr := st.UpdateView(resp.Full_tree_head, resp.Search /*@, p/2 @*/)
-	if updateErr != nil {
-		resultErr = updateErr
-		determined = true
+	// Read Prev_greatest while resp.Inv() is unfolded
+	start := uint32(0)
+	if resp.Prev_greatest != nil {
+		start = *resp.Prev_greatest + 1
 	}
+	end := int(resp.New_version)
 
-	// Copy timestamps while resp.Search.Inv() is still accessible
-	//@ unfold acc(resp.Search.Inv(), p/2)
-	timestamps := make([]uint64, len(resp.Search.Timestamps))
-	copy(timestamps, resp.Search.Timestamps /*@, p/4 @*/)
-	//@ fold acc(resp.Search.Inv(), p/2)
-
-	// Phase 2: Build prefix trees
 	//@ fold acc(resp.Inv(), p)
 
-	var trees []*proofs.PrefixTree
-	var rootHashes []*[sha256.Size]byte
-	if !determined {
-		var buildErr error
-		trees, rootHashes, buildErr = buildMonitorPrefixTrees(resp, n /*@, p @*/)
-		if buildErr != nil {
-			resultErr = buildErr
-			determined = true
-		}
-	}
-
-	// Phase 3: Per-entry direct path inspection (Section 8.2)
-	new_map = make([]client.MonitoringMapEntry, 0)
-	if !determined {
-		search_tree := client.MkImplicitBinarySearchTree(size)
-		distinguished := client.ComputeDistinguishedSet(search_tree, timestamps, config.ReasonableMonitoringWindow /*@, 1/2 @*/)
-		//@ assert size <= uint64(len(trees))
-
-		// Process entries
-		//@ ghost var versions seq[uint32] = seq[uint32]{}
-		//@ ghost var positions seq[uint64] = seq[uint64]{}
-
-		//@ invariant acc(monitor_map)
-		//@ invariant acc(new_map)
-		//@ invariant acc(trees)
-		//@ invariant acc(rootHashes)
-		//@ invariant acc(label, p)
-		//@ invariant acc(config, p)
-		//@ invariant search_tree != nil && acc(search_tree.Inv(), 1/4)
-		//@ invariant acc(distinguished)
-		//@ invariant low(len(distinguished))
-		//@ invariant forall j int :: 0 <= j && j < len(distinguished) ==> low(distinguished[j])
-		//@ invariant size <= uint64(len(trees))
-		//@ invariant forall i int :: i >= 0 && i < len(trees) ==> acc(&trees[i])
-		//@ invariant forall i int :: {&trees[i]} i >= 0 && i < len(trees) ==> acc(trees[i])
-		//@ invariant forall i int :: i >= 0 && i < len(trees) ==> trees[i] != nil
-		//@ invariant 0 <= mIdx && mIdx <= len(monitor_map)
-		//@ invariant low(mIdx)
-		//@ invariant !determined ==> resultErr == nil
-		//@ invariant low(versions)
-		//@ invariant len(versions) == len(new_map)
-		//@ invariant forall j int :: 0 <= j && j < len(new_map) ==> new_map[j].Version == versions[j]
-		//@ invariant low(positions)
-		//@ invariant len(positions) == len(new_map)
-		//@ invariant forall j int :: 0 <= j && j < len(new_map) ==> new_map[j].Position == positions[j]
-		//@ invariant len(new_map) == mIdx
-		//@ invariant low(len(new_map))
-		//@ invariant forall j int :: 0 <= j && j < len(monitor_map) ==> low(monitor_map[j].Version)
-		//@ invariant forall j int :: 0 <= j && j < len(monitor_map) ==> low(monitor_map[j].Position)
-		for mIdx := 0; mIdx < len(monitor_map); mIdx++ {
-			entry := monitor_map[mIdx]
-			//@ assert low(entry.Version)
-			//@ assert low(entry.Position)
-			newPosition := entry.Position
-
-			// All distinguished access is OUTSIDE if !determined so both
-			// executions touch distinguished identically (preserves low() in hyper mode)
-			isDistinguished := false
-			entryPosInt := int(entry.Position)
-			if entryPosInt >= 0 && entryPosInt < len(distinguished) {
-				isDistinguished = distinguished[entryPosInt]
-			}
-
-			// Compute pathRight unconditionally when !isDistinguished
-			// (isDistinguished is low(), so both executions take the same branch)
-			pathRight := make([]uint64, 0)
-			if !isDistinguished {
-				pathRight = client.DirectPathRight(search_tree, entry.Position, distinguished /*@, 1/8 @*/)
-			}
-
-			tVal := uint64(entry.Version)
-			//@ invariant acc(pathRight)
-			//@ invariant acc(trees)
-			//@ invariant acc(rootHashes)
-			//@ invariant acc(label, p)
-			//@ invariant acc(search_tree.Inv(), 1/4)
-			//@ invariant acc(distinguished)
-
-			//Distinguished nodes are low as they are public (Otherwise it doesn't make sense to make them as checkpoints)
-
-			//@ invariant low(len(distinguished))
-			//@ invariant forall j int :: 0 <= j && j < len(distinguished) ==> low(distinguished[j])
-			//@ invariant size <= uint64(len(trees))
-			//@ invariant forall i int :: i >= 0 && i < len(trees) ==> acc(&trees[i])
-			//@ invariant forall i int :: {&trees[i]} i >= 0 && i < len(trees) ==> acc(trees[i])
-			//@ invariant forall i int :: i >= 0 && i < len(trees) ==> trees[i] != nil
-			//@ invariant 0 <= pIdx && pIdx <= len(pathRight)
-			//@ invariant low(pIdx)
-			//@ invariant low(len(pathRight))
-			//@ invariant forall j int :: 0 <= j && j < len(pathRight) ==> low(pathRight[j])
-			//@ invariant low(newPosition)
-			//@ invariant !determined ==> resultErr == nil
-			for pIdx := 0; pIdx < len(pathRight); pIdx++ {
-				pos := pathRight[pIdx]
-				posIdx := int(pos)
-				if !determined {
-					if posIdx < 0 || posIdx >= len(trees) {
-						resultErr = errors.New("monitoring check failed: position out of bounds")
-						determined = true
-					} else {
-						Prefix_tree := trees[posIdx]
-						rootHash := rootHashes[posIdx]
-						if Prefix_tree != nil {
-							steps /*@, tStarIdx @*/ := client.FullBinaryLadderSteps_with_tstar(tVal)
-
-							//@ ghost var labelSeq seq[byte] = getContent(label)
-							//@ ghost var rootHashSeq seq[byte] = getContent(rootHash[:])
-
-							cgRes, cgErr := client.CheckGreatest(Prefix_tree, steps, label, tVal, rootHash[:], size /*@, tStarIdx, labelSeq, rootHashSeq @*/)
-							if cgErr != nil {
-								resultErr = cgErr
-								determined = true
-							} else if cgRes == -1 {
-								// Hole found: version not in log
-								resultErr = errors.New("monitoring check failed: version not included")
-								determined = true
-							}
-						}
-					}
+	// Verify each new version exists in the current log.
+	// Loop from 0 to end (both low) so both hyper-mode executions iterate
+	// the same number of times. Versions below start are skipped via inRange.
+	// Prefix trees are rebuilt each iteration because VerifyUpdateKey/CheckGreatest
+	// consume tree invariants and do not return them.
+	//@ invariant acc(resp.Inv(), p)
+	//@ invariant acc(label, p) && acc(config, p)
+	//@ invariant label != nil
+	//@ invariant low(len(label)) && (forall i int :: 0 <= i && i < len(label) ==> low(label[i]))
+	//@ invariant !determined ==> resultErr == nil
+	//@ invariant low(vIdx)
+	//@ invariant 0 <= vIdx && vIdx <= end + 1
+	//@ invariant low(size)
+	//@ invariant size > 0
+	//@ invariant size <= uint64(n)
+	for vIdx := 0; vIdx <= end; vIdx++ {
+		v := uint32(vIdx)
+		inRange := v >= start
+		if !determined && inRange {
+			trees, rootHashes, buildErr := buildUpdatePrefixTrees(resp, n /*@, p @*/)
+			if buildErr != nil {
+				resultErr = buildErr
+				determined = true
+			} else {
+				prev := uint32(0)
+				if v > 0 {
+					prev = v - 1
 				}
-				// Always advance position — both executions stay in sync
-				newPosition = pos
+				//@ assert size <= uint64(len(trees))
+				ok, verifyErr := VerifyUpdateKey(trees, rootHashes, size, label, v, prev, config /*@, p/4 @*/)
+				if verifyErr != nil {
+					resultErr = verifyErr
+					determined = true
+				} else if !ok {
+					resultErr = errors.New("history verification failed: version not found in log")
+					determined = true
+				}
 			}
-
-			// Always append to keep new_map/versions in sync across executions
-			newEntry := client.MonitoringMapEntry{
-				Position: newPosition,
-				Version:  entry.Version,
-			}
-			new_map = append( /*@ perm(1/2), @*/ new_map, newEntry)
-			//@ ghost versions = versions ++ seq[uint32]{entry.Version}
-			//@ ghost positions = positions ++ seq[uint64]{newPosition}
 		}
 	}
 
@@ -650,5 +440,5 @@ func VerifyMonitor(st *client.UserState, label []byte, resp MonitorResponse, mon
 	} else {
 		err = nil
 	}
-	return new_map, err
+	return err
 }
